// packages/template/src/store/generated/tendermint/spn/tendermint.spn.campaign/module/rest.ts
export interface CampaignQueryAllCampaignResponse {
  campaign?: CampaignCampaign[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface CampaignQueryAllMainnetAccountResponse {
  mainnetAccount?: CampaignMainnetAccount[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface CampaignQueryAllMainnetVestingAccountResponse {
  mainnetVestingAccount?: CampaignMainnetVestingAccount[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface CampaignQueryGetCampaignChainsResponse {
  campaignChains?: CampaignCampaignChains;
}

export interface CampaignQueryGetCampaignResponse {
  campaign?: CampaignCampaign;
}

export interface CampaignQueryGetMainnetAccountResponse {
  mainnetAccount?: CampaignMainnetAccount;
}

export interface CampaignQueryGetMainnetVestingAccountResponse {
  mainnetVestingAccount?: CampaignMainnetVestingAccount;
}

export interface CampaignShareDelayedVesting {
  totalShares?: V1Beta1Coin[];
  vesting?: V1Beta1Coin[];

  /** @format int64 */
  endTime?: string;
}

export interface CampaignShareVestingOptions {
  delayedVesting?: CampaignShareDelayedVesting;
}

export interface ProtobufAny {
  "@type"?: string;
}

export interface RpcStatus {
  /** @format int32 */
  code?: number;
  message?: string;
  details?: ProtobufAny[];
}
// ...
export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;

export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: keyof Omit<Body, "body" | "bodyUsed">;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;

export interface ApiConfig<SecurityDataType = unknown> {
  baseUrl?: string;
  baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">;
  securityWorker?: (securityData: SecurityDataType) => RequestParams | void;
}

export interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
  data: D;
  error: E;
}

type CancelToken = Symbol | string | number;

export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
}

export class HttpClient<SecurityDataType = unknown> {
  public baseUrl: string = "";
  private securityData: SecurityDataType = null as any;
  private securityWorker: null | ApiConfig<SecurityDataType>["securityWorker"] = null;
  private abortControllers = new Map<CancelToken, AbortController>();

  private baseApiParams: RequestParams = {
    credentials: "same-origin",
    headers: {},
    redirect: "follow",
    referrerPolicy: "no-referrer",
  };

  constructor(apiConfig: ApiConfig<SecurityDataType> = {}) {
    Object.assign(this, apiConfig);
  }

  public setSecurityData = (data: SecurityDataType) => {
    this.securityData = data;
  };

  private addQueryParam(query: QueryParamsType, key: string) {
    const value = query[key];

    return (
      encodeURIComponent(key) +
      "=" +
      encodeURIComponent(Array.isArray(value) ? value.join(",") : typeof value === "number" ? value : `${value}`)
    );
  }

  protected toQueryString(rawQuery?: QueryParamsType): string {
    const query = rawQuery || {};
    const keys = Object.keys(query).filter((key) => "undefined" !== typeof query[key]);
    return keys
      .map((key) =>
        typeof query[key] === "object" && !Array.isArray(query[key])
          ? this.toQueryString(query[key] as QueryParamsType)
          : this.addQueryParam(query, key),
      )
      .join("&");
  }

  protected addQueryParams(rawQuery?: QueryParamsType): string {
    const queryString = this.toQueryString(rawQuery);
    return queryString ? `?${queryString}` : "";
  }

  private contentFormatters: Record<ContentType, (input: any) => any> = {
    [ContentType.Json]: (input: any) =>
      input !== null && (typeof input === "object" || typeof input === "string") ? JSON.stringify(input) : input,
    [ContentType.FormData]: (input: any) =>
      Object.keys(input || {}).reduce((data, key) => {
        data.append(key, input[key]);
        return data;
      }, new FormData()),
    [ContentType.UrlEncoded]: (input: any) => this.toQueryString(input),
  };

  private mergeRequestParams(params1: RequestParams, params2?: RequestParams): RequestParams {
    return {
      ...this.baseApiParams,
      ...params1,
      ...(params2 || {}),
      headers: {
        ...(this.baseApiParams.headers || {}),
        ...(params1.headers || {}),
        ...((params2 && params2.headers) || {}),
      },
    };
  }

  private createAbortSignal = (cancelToken: CancelToken): AbortSignal | undefined => {
    if (this.abortControllers.has(cancelToken)) {
      const abortController = this.abortControllers.get(cancelToken);
      if (abortController) {
        return abortController.signal;
      }
      return void 0;
    }

    const abortController = new AbortController();
    this.abortControllers.set(cancelToken, abortController);
    return abortController.signal;
  };

  public abortRequest = (cancelToken: CancelToken) => {
    const abortController = this.abortControllers.get(cancelToken);

    if (abortController) {
      abortController.abort();
      this.abortControllers.delete(cancelToken);
    }
  };

  public request = <T = any, E = any>({
    body,
    secure,
    path,
    type,
    query,
    format = "json",
    baseUrl,
    cancelToken,
    ...params
  }: FullRequestParams): Promise<HttpResponse<T, E>> => {
    const secureParams = (secure && this.securityWorker && this.securityWorker(this.securityData)) || {};
    const requestParams = this.mergeRequestParams(params, secureParams);
    const queryString = query && this.toQueryString(query);
    const payloadFormatter = this.contentFormatters[type || ContentType.Json];

    return fetch(`${baseUrl || this.baseUrl || ""}${path}${queryString ? `?${queryString}` : ""}`, {
      ...requestParams,
      headers: {
        ...(type && type !== ContentType.FormData ? { "Content-Type": type } : {}),
        ...(requestParams.headers || {}),
      },
      signal: cancelToken ? this.createAbortSignal(cancelToken) : void 0,
      body: typeof body === "undefined" || body === null ? null : payloadFormatter(body),
    }).then(async (response) => {
      const r = response as HttpResponse<T, E>;
      r.data = (null as unknown) as T;
      r.error = (null as unknown) as E;

      const data = await response[format]()
        .then((data) => {
          if (r.ok) {
            r.data = data;
          } else {
            r.error = data;
          }
          return r;
        })
        .catch((e) => {
          r.error = e;
          return r;
        });

      if (cancelToken) {
        this.abortControllers.delete(cancelToken);
      }

      if (!response.ok) throw data;
      return data;
    });
  };
}

/**
 * @title campaign/campaign.proto
 * @version version not set
 */
export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
  /**
   * No description
   *
   * @tags Query
   * @name QueryCampaignAll
   * @summary Queries a list of campaign items.
   * @request GET:/tendermint/spn/campaign/campaign
   */
  queryCampaignAll = (
    query?: {
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.countTotal"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<CampaignQueryAllCampaignResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/campaign`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryCampaign
   * @summary Queries a campaign by id.
   * @request GET:/tendermint/spn/campaign/campaign/{campaignID}
   */
  queryCampaign = (campaignID: string, params: RequestParams = {}) =>
    this.request<CampaignQueryGetCampaignResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/campaign/${campaignID}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryCampaignChains
   * @summary Queries a campaignChains by index.
   * @request GET:/tendermint/spn/campaign/campaign_chains/{campaignID}
   */
  queryCampaignChains = (campaignID: string, params: RequestParams = {}) =>
    this.request<CampaignQueryGetCampaignChainsResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/campaign_chains/${campaignID}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryMainnetAccountAll
   * @summary Queries a list of mainnetAccount items.
   * @request GET:/tendermint/spn/campaign/mainnet_account
   */
  queryMainnetAccountAll = (
    query?: {
      campaignID?: string;
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.countTotal"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<CampaignQueryAllMainnetAccountResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/mainnet_account`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryMainnetAccount
   * @summary Queries a mainnetAccount by index.
   * @request GET:/tendermint/spn/campaign/mainnet_account/{campaignID}/{address}
   */
  queryMainnetAccount = (campaignID: string, address: string, params: RequestParams = {}) =>
    this.request<CampaignQueryGetMainnetAccountResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/mainnet_account/${campaignID}/${address}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryMainnetVestingAccountAll
   * @summary Queries a list of mainnetVestingAccount items.
   * @request GET:/tendermint/spn/campaign/mainnet_vesting_account
   */
  queryMainnetVestingAccountAll = (
    query?: {
      campaignID?: string;
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.countTotal"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<CampaignQueryAllMainnetVestingAccountResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/mainnet_vesting_account`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryMainnetVestingAccount
   * @summary Queries a mainnetVestingAccount by index.
   * @request GET:/tendermint/spn/campaign/mainnet_vesting_account/{campaignID}/{address}
   */
  queryMainnetVestingAccount = (campaignID: string, address: string, params: RequestParams = {}) =>
    this.request<CampaignQueryGetMainnetVestingAccountResponse, RpcStatus>({
      path: `/tendermint/spn/campaign/mainnet_vesting_account/${campaignID}/${address}`,
      method: "GET",
      format: "json",
      ...params,
    });
}